Move workbook generation out, updated model
